OpenBCI-Csound/Experiments/requirements.txt
python-osc>=1.8.0
numpy>=1.21.0
scipy>=1.7.0
brainflow>=5.0.0
pyserial>=3.5